renaming and adding pylot pathing
