+++ src/RRTStar/rrt_star.cpp
@@ -1,4 +1,4 @@
-#include "rrt.h"
+#include "rrt_star.h"
 
 #include <algorithm>
 #include <vector>
